Small changes to transaction table, import data from CSV.

+++ EveWarehouse.Database/Live/Tables/InventoryEntries.sql
@@ -1,7 +1,7 @@
 ﻿CREATE TABLE [Live].[InventoryEntries]
 (
     [Id] BIGINT NOT NULL IDENTITY PRIMARY KEY, 
-    [ItemId] BIGINT NOT NULL, 
+    [ItemId] INT NOT NULL, 
     [Price] DECIMAL(27, 2) NOT NULL, 
     [Movement] BIGINT NOT NULL,
     [Date] DATETIME2 NOT NULL,
